% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot_sample.R
\name{Boot_sample}
\alias{Boot_sample}
\title{Function to bootstrap the data}
\usage{
Boot_sample(data, strata, N_Sim = 1000)
}
\arguments{
\item{data}{A data.frame that is going to be bootstrapped.}

\item{strata}{The name of the column that indicates the stratification groups.}

\item{N_Sim}{Number of bootstraps.}
}
\description{
Function to bootstrap the data
}
