% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_CI.R
\name{bs_CI}
\alias{bs_CI}
\title{Function to calculate bootstrap Confidence Intervals}
\usage{
bs_CI(data, strata, N_Sim, full = F)
}
\arguments{
\item{data}{A data.frame that is going to be bootstrapped.}

\item{strata}{The name of the column that indicates the stratification groups.}

\item{N_Sim}{Number of bootstraps.}

\item{full}{if False only CIs are returned, if T the data from bootstraps is also returned}
}
\description{
Function to calculate bootstrap Confidence Intervals
}
