% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_reg.R
\name{bs_reg}
\alias{bs_reg}
\title{Function to calculate regression from the bootstraps}
\usage{
bs_reg(Bs, formula)
}
\arguments{
\item{Bs}{A list resulting from the function Boot_sample}

\item{formula}{The formula to use for the regression}
}
\description{
Function to calculate regression from the bootstraps
}
