% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEAC.R
\name{CEAC}
\alias{CEAC}
\title{Function to plot the CEAC}
\usage{
CEAC(data, strata, N_Sim = 1000, NB, full = F)
}
\arguments{
\item{data}{A data.frame containing the information of cost and effect.}

\item{strata}{The name of the column that indicates the stratification groups.}

\item{N_Sim}{Number of bootstraps.}

\item{NB}{a vector containing the values for the willingness to pay}

\item{full}{if False only a plot is returned, if T the data from bootstraps is also returned}
}
\description{
Function to plot the CEAC
}
